server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
vaderSentiment==3.3.2
google-generativeai==0.3.1
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
